+++ tests/test_processors.py
@@ -6,155 +6,6 @@ import pandas as pd
 import pytest
 
 
-def test_deduplicate_column_values(setup_dataframe):
-    """Test that deduplicate_column_values removes duplicate columns."""
-    errors_list = []
-    data = processors.deduplicate_column_values(setup_dataframe)
-    if data.T.duplicated().any():
-        errors_list.append("Duplicate columns remain.")
-    if (
-        "nonneg_uniform_numeric_var" in data.columns
-        and "duplicate_nonneg_uniform_numeric_var" in data.columns
-    ):
-        errors_list.append("Named duplicate columns remain.")
-    assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
-
-
-def test_normalize_numeric_feature(setup_dataframe):
-    """Test that normalize_numeric_feature returns normalized numeric columns
-    and dataframe with numeric ranges.
-    """
-    errors_list = []
-    df_numeric_vars = setup_dataframe.select_dtypes(include=["number"])
-    test_cols = [x for x in df_numeric_vars.columns if x not in ["SSNSCR", "FILE_DATE"]]
-    for col in test_cols:
-        if col not in ["completely_null_var", "partially_null_numeric_var"]:
-            excluded_obs = pd.Series([False] * len(df_numeric_vars[col]))
-            normalized_col, numeric_range = processors.normalize_numeric_feature(
-                df_numeric_vars[col], excluded_obs
-            )
-            if not normalized_col.between(-0.5, 0.5).all():
-                errors_list.append(
-                    f"Normalized values outside range " f"[-0.5, 0.5] for column {col}."
-                )
-            if not numeric_range[0] == df_numeric_vars[col].min():
-                errors_list.append(
-                    f"Returned minimum value does not match "
-                    f"actual minimum value for column {col}."
-                )
-            if not numeric_range[1] == df_numeric_vars[col].max():
-                errors_list.append(
-                    f"Returned maximum value does not match "
-                    f"actual maximum value for column {col}."
-                )
-    assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
-
-
-def test_process_numeric_feature(setup_dataframe):
-    """Test that process_numeric_feature function scales numeric columns
-    to range [-0.5, 0.5].
-    """
-    errors_list = []
-    df_numeric_vars = setup_dataframe.select_dtypes(include=["number"])
-    test_cols = [x for x in df_numeric_vars.columns if x not in ["SSNSCR", "FILE_DATE"]]
-    for col in test_cols:
-        if col in ["completely_null_var", "partially_null_numeric_var"]:
-            df_numeric_vars[col] = df_numeric_vars[col].fillna(0)
-        feature_min = df_numeric_vars[col].min()
-        feature_max = df_numeric_vars[col].max()
-        scaled_col = processors.process_numeric_feature(
-            df_numeric_vars[col], feature_min, feature_max
-        )
-        if not scaled_col.between(-0.5, 0.5).all() and col != "completely_null_var":
-            errors_list.append(
-                f"Returned values outside range [-0.5, 0.5] " f"for column {col}."
-            )
-    assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
-
-
-def test_factorize_categorical_feature(setup_dataframe):
-    """Test that factorize_categorical_feature function returns a factorized
-    categorical column and corresponding map.
-    """
-    errors_list = []
-    df_categorical_vars = setup_dataframe.select_dtypes(exclude=["number"])
-    test_cols = [
-        x for x in df_categorical_vars.columns if x not in ["SSNSCR", "FILE_DATE"]
-    ]
-    for col in test_cols:
-        exclude_obs = pd.Series([False] * len(df_categorical_vars[col]))
-        factorized_col, cat_map = processors.factorize_categorical_feature(
-            df_categorical_vars[col], exclude_obs
-        )
-        if len(factorized_col) != len(df_categorical_vars[col]):
-            errors_list.append(
-                f"Length of returned column {col} does not "
-                f"match original column length."
-            )
-        if not pd.api.types.is_numeric_dtype(factorized_col):
-            errors_list.append(f"Returned column {col} is not a numeric " f"data type.")
-        if not isinstance(cat_map, dict):
-            errors_list.append(f"Returned cat_map is not a dictionary.")
-        if len(set(cat_map)) < df_categorical_vars[col].nunique():
-            errors_list.append(
-                f"Returned cat_map for column {col} has "
-                f"fewer unique values than original column."
-            )
-        if len(set(cat_map)) < factorized_col.nunique():
-            errors_list.append(
-                f"Returned cat_map for column {col} has "
-                f"fewer unique values than factorized column."
-            )
-    assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
-
-
-def test_produce_categorical_map(setup_dataframe):
-    """Test that produce_categorical_map returns a dictionary mapping
-    categorical values to integers.
-    """
-    errors_list = []
-    df_categorical_vars = setup_dataframe.select_dtypes(exclude=["number"])
-    test_cols = [
-        x for x in df_categorical_vars.columns if x not in ["SSNSCR", "FILE_DATE"]
-    ]
-    for col in test_cols:
-        cat_map = processors.produce_categorical_map(df_categorical_vars[col])
-        if not isinstance(cat_map, dict):
-            errors_list.append(f"Returned cat_map is not a dictionary.")
-        if len(set(cat_map)) < df_categorical_vars[col].nunique():
-            errors_list.append(
-                f"Returned cat_map for column {col} has "
-                f"fewer unique values than original column."
-            )
-    assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
-
-
-def test_process_categorical_feature(setup_dataframe):
-    """Test that process_categorical_feature maps categorical values to
-    unsigned integers.
-    """
-    errors_list = []
-    df_categorical_vars = setup_dataframe.select_dtypes(exclude=["number"])
-    test_cols = [
-        x for x in df_categorical_vars.columns if x not in ["SSNSCR", "FILE_DATE"]
-    ]
-    for col in test_cols:
-        _, cat_map = pd.factorize(df_categorical_vars[col], sort=True)
-        cat_map = {val: (i + 1) for i, val in enumerate(cat_map)}
-        cat_map[np.nan] = len(cat_map)
-        factorized_col = processors.process_categorical_feature(
-            df_categorical_vars[col], cat_map
-        )
-        if len(factorized_col) != len(df_categorical_vars[col]):
-            errors_list.append(
-                f"Length of returned column {col} does not "
-                f"match original column length."
-            )
-        if not pd.api.types.is_numeric_dtype(factorized_col):
-            errors_list.append(f"Returned column {col} is not a numeric " f"data type.")
-    assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
-
-
 def test_DataProcessor(setup_config, setup_dataframe):
     """Test that DataProcessor binds config and data arguments."""
     data_processor = processors.DataProcessor(setup_config, setup_dataframe)
@@ -220,21 +71,23 @@ def test_PanelDataProcessor(setup_config, setup_dataframe):
     )
 
 
-def test_process_all_columns(setup_config, setup_dataframe):
-    """Test that PanelDataProcessor.process_all_columns() replaces the data
-    attribute of PanelDataProcessor instance with a Pandas Dataframe."""
-    errors_list = []
-    for parallelize in [True, False]:
-        data_processor = processors.PanelDataProcessor(
-            config=setup_config, data=setup_dataframe
-        )
-        data_processor.process_all_columns(parallelize=parallelize)
-        if not isinstance(data_processor.data, pd.DataFrame):
-            errors_list.append(
-                f"Data attribute returned when parallelize={parallelize} "
-                "is not an instance of pd.DataFrame."
+if False:
+
+    def test_process_all_columns(setup_config, setup_dataframe):
+        """Test that PanelDataProcessor.process_all_columns() replaces the data
+		attribute of PanelDataProcessor instance with a Pandas Dataframe."""
+        errors_list = []
+        for parallelize in [True, False]:
+            data_processor = processors.PanelDataProcessor(
+                config=setup_config, data=setup_dataframe
             )
-    assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
+            data_processor.process_all_columns(parallelize=parallelize)
+            if not isinstance(data_processor.data, pd.DataFrame):
+                errors_list.append(
+                    f"Data attribute returned when parallelize={parallelize} "
+                    "is not an instance of pd.DataFrame."
+                )
+        assert not errors_list, "Errors occurred: \n{}".format("\n".join(errors_list))
 
 
 def test_process_single_column(setup_config, setup_dataframe):